Use same format for 35 mm equivalent as for normal focal length
darcs-hash:20071006135746-8a255-886dc435f84d65a5888f2423f2b437ffe93087ee.gz

+++ lazygal/metadata.py
@@ -194,7 +194,7 @@ class ExifTags(pyexiv2.Image):
             val = str(self['Exif.Photo.FocalLength']).split('/')
             if len(val) == 1: val.append('1')
             foclength = float(val[0]) / float(val[1])
-            flen += ' (35 mm equivalent: %d mm)' % int(foclength / ccdwidth * 36 + 0.5)
+            flen += ' (35 mm equivalent: %01f mm)' % (foclength / ccdwidth * 36 + 0.5)
         except IndexError:
             return flen
